Add files via upload
Test prints for your new 3D printer. Start with the numbered ones, and slice them to the temperature labeled on the block. This will tell you the best temperature for a particular filament. repeat for each new filament

Once you have figured out the best temperature, go ahead and try printing one of the other test prints. NOTE: Benchy is the hardest one to get right, so save that one until you think you've got it as perfect as you can get it